pyproject.toml
[project]
name = "rraushan-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mkdocs-awesome-pages-plugin>=2.10.1",
    "mkdocs-material[blog]>=9.7.6",
]
